Revise all routes to use conversation/opinion in the route name
