+++ services/agora/src/pages/moderate/conversation/[postSlugId]/index.vue
@@ -15,7 +15,7 @@
   >
     <div class="container">
       <div class="title">
-        Moderate the opinion "{add opinion title excerpt}"
+        Moderate the conversation "{add conversation title excerpt}"
       </div>
 
       <q-select
@@ -62,32 +62,31 @@
 
 <script setup lang="ts">
 import { useBackendModerateApi } from "src/utils/api/moderation";
-import { useRoute } from "vue-router";
+import { useRoute, useRouter } from "vue-router";
 import { onMounted, ref } from "vue";
 import type {
-  ModerationActionComments,
+  ModerationActionPosts,
   ModerationReason,
 } from "src/shared/types/zod";
 import ZKButton from "src/components/ui-library/ZKButton.vue";
 import {
-  moderationActionCommentsMapping,
+  moderationActionPostsMapping,
   moderationReasonMapping,
 } from "src/utils/component/moderations";
+import { usePostStore } from "src/stores/post";
 import MainLayout from "src/layouts/MainLayout.vue";
 
-const {
-  moderateComment,
-  fetchCommentModeration,
-  cancelModerationCommentReport,
-} = useBackendModerateApi();
+const { moderatePost, fetchPostModeration, cancelModerationPostReport } =
+  useBackendModerateApi();
 
 const route = useRoute();
+const router = useRouter();
 
-const DEFAULT_MODERATION_ACTION = "move";
-const moderationAction = ref<ModerationActionComments>(
-  DEFAULT_MODERATION_ACTION
-);
-const actionMapping = ref(moderationActionCommentsMapping);
+const { loadPostData } = usePostStore();
+
+const DEFAULT_MODERATION_ACTION = "lock";
+const moderationAction = ref<ModerationActionPosts>(DEFAULT_MODERATION_ACTION);
+const actionMapping = ref(moderationActionPostsMapping);
 
 const DEFAULT_MODERATION_REASON = "misleading";
 const moderationReason = ref<ModerationReason>(DEFAULT_MODERATION_REASON);
@@ -97,12 +96,12 @@ const moderationExplanation = ref("");
 
 const hasExistingDecision = ref(false);
 
-let commentSlugId: string | null = null;
+let postSlugId: string | null = null;
 if (
-  route.name == "/moderate/comment/[commentSlugId]/" &&
-  typeof route.params.commentSlugId == "string"
+  route.name == "/moderate/conversation/[postSlugId]/" &&
+  typeof route.params.postSlugId == "string"
 ) {
-  commentSlugId = route.params.commentSlugId;
+  postSlugId = route.params.postSlugId;
 }
 
 onMounted(async () => {
@@ -110,8 +109,8 @@ onMounted(async () => {
 });
 
 async function initializeData() {
-  if (commentSlugId != null) {
-    const response = await fetchCommentModeration(commentSlugId);
+  if (postSlugId != null) {
+    const response = await fetchPostModeration(postSlugId);
     hasExistingDecision.value = response.status == "moderated";
     if (response.status == "moderated") {
       moderationAction.value = response.action;
@@ -123,38 +122,41 @@ async function initializeData() {
       moderationReason.value = DEFAULT_MODERATION_REASON;
     }
   } else {
-    console.log("Missing comment slug ID");
+    console.log("Missing post slug ID");
   }
 }
 
 async function clickedWithdraw() {
-  if (commentSlugId) {
-    await cancelModerationCommentReport(commentSlugId);
-    await initializeData();
-    // TODO: redirect to comment
-    // await router.push({
-    //   name: "single-post",
-    //   params: { postSlugId: postSlugId },
-    // });
+  if (postSlugId) {
+    const isSuccessful = await cancelModerationPostReport(postSlugId);
+    if (isSuccessful) {
+      await initializeData();
+      await loadPostData(false);
+      await router.push({
+        name: "/conversation/[postSlugId]",
+        params: { postSlugId: postSlugId },
+      });
+    }
   } else {
     console.log("Missing comment slug ID");
   }
 }
 
 async function clickedSubmit() {
-  if (commentSlugId) {
-    const isSuccessful = await moderateComment(
-      commentSlugId,
+  if (postSlugId) {
+    const isSuccessful = await moderatePost(
+      postSlugId,
       moderationAction.value,
       moderationReason.value,
       moderationExplanation.value
     );
+
     if (isSuccessful) {
-      // TODO: redirect to comment
-      // await router.push({
-      //   name: "single-post",
-      //   params: { postSlugId: postSlugId },
-      // });
+      await loadPostData(false);
+      await router.push({
+        name: "/conversation/[postSlugId]",
+        params: { postSlugId: postSlugId },
+      });
     }
   }
 }
